+++ apps/nextjs-app/src/backend/features/poem/SearchPoems/SearchPoemsQuery.ts
@@ -1,4 +1,4 @@
-import { HttpInternalServerError } from '@belgattitude/http-exception';
+import { HttpInternalServerError } from '@httpx/exception';
 import type { PrismaClientDbMain } from '@your-org/db-main-prisma';
 import type { UnPromisify } from '@your-org/ts-utils';
 import type { SearchPoemsParams } from './SearchPoems.types';