+++ apps/nextjs-app/src/pages/home.tsx
@@ -1,8 +1,8 @@
-import { HttpBadRequest } from '@belgattitude/http-exception';
+import { HttpBadRequest } from '@httpx/exception';
 import type { GetServerSideProps, InferGetServerSidePropsType } from 'next';
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { homeConfig } from '@/features/home/home.config';
 import { HomePage } from '@/features/home/pages';
+import { getServerTranslations } from '@/lib/i18n';
 
 type Props = {
   /** Add HomeRoute props here */
@@ -24,7 +24,7 @@ export const getServerSideProps: GetServerSideProps<Props> = async (
   const { i18nNamespaces } = homeConfig;
   return {
     props: {
-      ...(await serverSideTranslations(locale, i18nNamespaces)),
+      ...(await getServerTranslations(locale, i18nNamespaces)),
     },
   };
 };