+++ packages/db-main-prisma/src/index.ts
@@ -1,3 +1,5 @@
 export { PrismaClient as PrismaClientDbMain } from '@prisma/client';
+export type { Prisma as PrismaDbMain } from '@prisma/client';
 export { PrismaManager } from './prisma-manager';
 export type { PrismaClientOptions } from './prisma-manager';
+export type { default as DbMainPrismaTypes } from './generated/prisma-pothos-types';