+++ packages/db-main-prisma/src/seeds/index.ts
@@ -1,0 +1,2 @@
+export { UserSeeds } from './UserSeeds';
+export { PoemSeeds } from './PoemSeeds';