+++ apps/nextjs-app/CHANGELOG.md
@@ -1,5 +1,39 @@
 # nextjs-app
 
+## 3.58.0
+
+### Minor Changes
+
+- [#2941](https://github.com/belgattitude/nextjs-monorepo-example/pull/2941) [`e0b79462`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e0b7946254b8b65a8b2bf0932350753397236ea8) Thanks [@belgattitude](https://github.com/belgattitude)! - Initial support for @nextvalid/zod-request
+
+  https://github.com/belgattitude/nextvalid/tree/main/packages/zod-request#install
+
+### Patch Changes
+
+- [#2941](https://github.com/belgattitude/nextjs-monorepo-example/pull/2941) [`e0b79462`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e0b7946254b8b65a8b2bf0932350753397236ea8) Thanks [@belgattitude](https://github.com/belgattitude)! - Use https://belgattitude.github.io/httpx / https://github.com/belgattitude/httpx
+
+- Updated dependencies [[`e0b79462`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e0b7946254b8b65a8b2bf0932350753397236ea8)]:
+  - @your-org/core-lib@3.15.2
+
+## 3.57.0
+
+### Minor Changes
+
+- [#2932](https://github.com/belgattitude/nextjs-monorepo-example/pull/2932) [`f7507455`](https://github.com/belgattitude/nextjs-monorepo-example/commit/f7507455f2ad5b9f4496a758651007557971a694) Thanks [@belgattitude](https://github.com/belgattitude)! - Upgrade to next-i18next v13.0.0
+
+  https://github.com/i18next/next-i18next/releases/tag/v13.0.0
+
+## 3.56.0
+
+### Minor Changes
+
+- [#2862](https://github.com/belgattitude/nextjs-monorepo-example/pull/2862) [`fab08286`](https://github.com/belgattitude/nextjs-monorepo-example/commit/fab08286695522571506f4a65db7f8683f3a09bd) Thanks [@belgattitude](https://github.com/belgattitude)! - Move to graphql-yoga from apollo and micro
+
+### Patch Changes
+
+- Updated dependencies [[`fab08286`](https://github.com/belgattitude/nextjs-monorepo-example/commit/fab08286695522571506f4a65db7f8683f3a09bd)]:
+  - @your-org/db-main-prisma@2.13.0
+
 ## 3.55.0
 
 ### Minor Changes